% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_selected.R
\name{buffer_selected}
\alias{buffer_selected}
\title{Buffer selected features}
\usage{
buffer_selected(data, var, threshold, bufdist)
}
\arguments{
\item{data}{the dataset}

\item{var}{the attribute/variable to select from}

\item{threshold}{the value to be met by the condition}

\item{bufdist}{the buffer distance to create around selected features}
}
\value{
a buffer of a certain radius around selected features in a dataset
}
\description{
Buffer selected features
}
\examples{
translines <- sf::st_read(system.file("extdata/transmissionLines", "transmissionLine.shp",
 package="siteselection"))
buffer_selected(translines, "VOLTAGE", 400, 10)

}
