% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_selected.R
\name{clip_selected}
\alias{clip_selected}
\title{Clip selected features}
\usage{
clip_selected(data1, data2, d1var, d2var, d1threshold, d2threshold)
}
\arguments{
\item{data1}{input dataset}

\item{data2}{clip dataset}

\item{d1var}{variable for input dataset}

\item{d2var}{variable for clip dataset}

\item{d1threshold}{condition to be met in d1var}

\item{d2threshold}{condition to be met in d2var}
}
\value{
clippped features that meet condition in both datasets
}
\description{
Clip selected features
}
\examples{
counties <- sf::st_read(system.file("extdata/counties", "counties.shp",
package="siteselection"))
windpw <- sf::st_read(system.file("extdata/windPower", "windPower.shp",
package="siteselection"))
clip_selected(counties, windpw, "Population","WindPowerClass", 20000, 3 )
}
